Using webpack to try to split files but maps const in not working now.
