+++ node/risk-app/server/views/map.html
@@ -63,7 +63,7 @@
 <script>
   window.mapCategories = {{ maps | dump | safe }}
 </script>
-<script src="{{ assetPath }}/js/map-page.js"></script>
+<script src="{{ assetPath }}/js/map-page/map-page.js"></script>
 
 {% endblock %}
